[Refactor] [core] Combine unnecessary packages
