+++ seatunnel-core/seatunnel-core-sql/src/test/java/org/apache/seatunnel/core/sql/SqlVariableSubstitutionTest.java
@@ -18,7 +18,6 @@
 package org.apache.seatunnel.core.sql;
 
 import org.apache.seatunnel.command.FlinkCommandArgs;
-import org.apache.seatunnel.core.sql.job.JobInfo;
 import org.apache.seatunnel.utils.CommandLineUtils;
 
 import org.apache.commons.io.FileUtils;
